Recover cursor state when editor window loses focus
The show count [1] of the mouse cursor, controlled by `ShowCursor()`, is
thread-local and has an effect only when the cursor is above a window of the
same thread. On the other hand, `GetCursorInfo()` reports the asynchronous
state of the cursor, which may be outside the editor window and doesn't
reflect the show count of our thread.

So, it is incorrect to manipulate the show count according to the result of
`GetCursorInfo()`. This commit reverts relevant changes about `cursorIsHidden`
in upstream Scintilla.

When the editor window loses focus but the cursor is not above it, the system
does not send a `WM_MOUSELEAVE` notification. So we also have to recover the
cursor state in `WM_KILLFOCUS`.

This closes https://github.com/rizonesoft/Notepad3/issues/5369

[1] https://devblogs.microsoft.com/oldnewthing/20091217-00/?p=15643

Signed-off-by: LIU Hao <lh_mouse@126.com>